feat(core): extract pure utilities and JobManager (phase-3 PR1)
Move pure helper functions out of WikiMemory into utils/pure.ts.
Centralize all mutex locking and status broadcasting in a new
JobManager class; WikiMemory delegates all lock operations through
jobManager.acquireLock/releaseLock rather than managing raw Sets.

Export HOOK_TIMEOUT_MARKER so PR2 MaintenanceService can reference it.
Update test helpers to access state via jobManager internals.
