+++ packages/core/__tests__/jobs.test.ts
@@ -47,7 +47,7 @@ describe('job mutex', () => {
 
   it('runLibrarian throws WikiBusyError when prune is running for same entity', async () => {
     const wiki = await freshWiki(slowProvider(0));
-    (wiki as any).activeMaintenanceJobs.add('jobs_:e1:prune');
+    (wiki as any).jobManager.activeMaintenanceJobs.add('jobs_:e1:prune');
     const err = await wiki.runLibrarian('e1').catch((e) => e);
     expect(err).toBeInstanceOf(WikiBusyError);
     expect(err.operation).toBe('prune');
@@ -55,7 +55,7 @@ describe('job mutex', () => {
 
   it('runHeal throws WikiBusyError when prune is running for same entity', async () => {
     const wiki = await freshWiki(slowProvider(0));
-    (wiki as any).activeMaintenanceJobs.add('jobs_:e1:prune');
+    (wiki as any).jobManager.activeMaintenanceJobs.add('jobs_:e1:prune');
     const err = await wiki.runHeal('e1').catch((e) => e);
     expect(err).toBeInstanceOf(WikiBusyError);
     expect(err.operation).toBe('prune');
@@ -63,7 +63,7 @@ describe('job mutex', () => {
 
   it('ingestDocument throws WikiBusyError when prune is running for same entity', async () => {
     const wiki = await freshWiki(slowProvider(0));
-    (wiki as any).activeMaintenanceJobs.add('jobs_:e1:prune');
+    (wiki as any).jobManager.activeMaintenanceJobs.add('jobs_:e1:prune');
     const sourceHash = 'a'.repeat(64);
     const err = await wiki.ingestDocument('e1', { sourceRef: 'doc1', sourceHash, documentChunk: 'some content here' }).catch((e) => e);
     expect(err).toBeInstanceOf(WikiBusyError);
@@ -72,7 +72,7 @@ describe('job mutex', () => {
 
   it('prune guard does not bleed to different entity', async () => {
     const wiki = await freshWiki(slowProvider(100));
-    (wiki as any).activeMaintenanceJobs.add('jobs_:e1:prune');
+    (wiki as any).jobManager.activeMaintenanceJobs.add('jobs_:e1:prune');
     // e2 should not be affected by e1 prune lock
     const p = wiki.runLibrarian('e2');
     await expect(p).resolves.toBeUndefined();
